+++ src/main/java/com/sion/concertbooking/intefaces/presentation/rest/PaymentController.java
@@ -3,9 +3,9 @@ package com.sion.concertbooking.intefaces.presentation.rest;
 import com.sion.concertbooking.application.payment.PaymentFacade;
 import com.sion.concertbooking.application.payment.PaymentCriteria;
 import com.sion.concertbooking.application.payment.PaymentResult;
-import com.sion.concertbooking.infrastructure.aspect.TokenInfo;
-import com.sion.concertbooking.infrastructure.aspect.TokenRequired;
-import com.sion.concertbooking.infrastructure.aspect.TokenUtils;
+import com.sion.concertbooking.intefaces.aspect.TokenInfo;
+import com.sion.concertbooking.intefaces.aspect.TokenRequired;
+import com.sion.concertbooking.intefaces.aspect.TokenUtils;
 import com.sion.concertbooking.intefaces.presentation.request.PaymentRequest;
 import com.sion.concertbooking.intefaces.presentation.response.PaymentResponse;
 import io.swagger.v3.oas.annotations.Operation;
