+++ src/main/java/com/sion/concertbooking/intefaces/presentation/rest/WaitingQueueController.java
@@ -3,9 +3,9 @@ package com.sion.concertbooking.intefaces.presentation.rest;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueInfo;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueDetailInfo;
 import com.sion.concertbooking.domain.watingqueue.WaitingQueueService;
-import com.sion.concertbooking.infrastructure.aspect.TokenInfo;
-import com.sion.concertbooking.infrastructure.aspect.TokenRequired;
-import com.sion.concertbooking.infrastructure.aspect.TokenUtils;
+import com.sion.concertbooking.intefaces.aspect.TokenInfo;
+import com.sion.concertbooking.intefaces.aspect.TokenRequired;
+import com.sion.concertbooking.intefaces.aspect.TokenUtils;
 import com.sion.concertbooking.intefaces.presentation.request.WaitingQueueRegisterRequest;
 import com.sion.concertbooking.intefaces.presentation.response.WaitingQueueInfoResponse;
 import com.sion.concertbooking.intefaces.presentation.response.WaitingQueueRegisterResponse;
